// frontend/src/api/generated/model/agentUpdate.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Mission Control API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentUpdateHeartbeatConfig } from "./agentUpdateHeartbeatConfig";
import type { AgentUpdateIdentityProfile } from "./agentUpdateIdentityProfile";

/**
 * Payload for patching an existing agent.
 */
export interface AgentUpdate {
  board_id?: string | null;
  is_gateway_main?: boolean | null;
  name?: string | null;
  status?: string | null;
  heartbeat_config?: AgentUpdateHeartbeatConfig;
  identity_profile?: AgentUpdateIdentityProfile;
  identity_template?: string | null;
  soul_template?: string | null;
}